[REFACTORING] 리뷰 반영
