+++ user-api/src/main/java/com/biengual/userapi/bookmark/domain/BookmarkStore.java
@@ -3,7 +3,7 @@ package com.biengual.userapi.bookmark.domain;
 public interface BookmarkStore {
 	void deleteBookmark(Long userId, Long bookmarkId);
 
-	void saveBookmark(BookmarkCommand.Create command, String detail, Double startTime);
+	void saveBookmark(BookmarkCommand.Create command);
 
 	BookmarkInfo.Position updateBookmark(BookmarkCommand.Update command);
 }
