+++ user-api/src/main/java/com/biengual/userapi/bookmark/infrastructure/BookmarkStoreImpl.java
@@ -1,9 +1,12 @@
 package com.biengual.userapi.bookmark.infrastructure;
 
 import static com.biengual.userapi.message.error.code.BookmarkErrorCode.*;
+import static com.biengual.userapi.message.error.code.ContentErrorCode.*;
 import static com.biengual.userapi.message.error.code.UserErrorCode.*;
 
 import java.util.Objects;
+
+import org.bson.types.ObjectId;
 
 import com.biengual.userapi.annotation.DataProvider;
 import com.biengual.userapi.bookmark.domain.BookmarkCommand;
@@ -11,7 +14,12 @@ import com.biengual.userapi.bookmark.domain.BookmarkEntity;
 import com.biengual.userapi.bookmark.domain.BookmarkInfo;
 import com.biengual.userapi.bookmark.domain.BookmarkRepository;
 import com.biengual.userapi.bookmark.domain.BookmarkStore;
+import com.biengual.userapi.content.domain.entity.ContentDocument;
+import com.biengual.userapi.content.domain.enums.ContentType;
+import com.biengual.userapi.content.repository.ContentRepository;
+import com.biengual.userapi.content.repository.ContentScriptRepository;
 import com.biengual.userapi.message.error.exception.CommonException;
+import com.biengual.userapi.script.domain.entity.YoutubeScript;
 import com.biengual.userapi.user.domain.entity.UserEntity;
 import com.biengual.userapi.user.repository.UserRepository;
 
@@ -22,6 +30,8 @@ import lombok.RequiredArgsConstructor;
 public class BookmarkStoreImpl implements BookmarkStore {
 	private final BookmarkRepository bookmarkRepository;
 	private final UserRepository userRepository;
+	private final ContentRepository contentRepository;
+	private final ContentScriptRepository contentScriptRepository;
 
 	@Override
 	public void deleteBookmark(Long userId, Long bookmarkId) {
@@ -29,11 +39,18 @@ public class BookmarkStoreImpl implements BookmarkStore {
 	}
 
 	@Override
-	public void saveBookmark(BookmarkCommand.Create command, String detail, Double startTime) {
+	public void saveBookmark(BookmarkCommand.Create command) {
+		// TODO: user -> command.userId() 로 수정 필요함, updateUserBookmark 로직도 수정 필요
 		UserEntity user = userRepository.findById(command.userId())
 			.orElseThrow(() -> new CommonException(USER_NOT_FOUND));
 
-		BookmarkEntity bookmark = command.toEntity(detail, startTime);
+		ContentDocument content = contentScriptRepository.findContentDocumentById(
+			new ObjectId(contentRepository.findMongoIdByContentId(command.contentId()))
+		).orElseThrow(() -> new CommonException(CONTENT_NOT_FOUND));
+
+		BookmarkEntity bookmark
+			= command.toEntity(extractDetail(command, content), extractStartTime(command, content));
+
 		bookmarkRepository.save(bookmark);
 		user.updateUserBookmark(bookmark);
 	}
@@ -52,5 +69,29 @@ public class BookmarkStoreImpl implements BookmarkStore {
 		bookmark.updateDescription(command.description());
 
 		return BookmarkInfo.Position.of(bookmark);
+	}
+
+	// Internal Methods ------------------------------------------------------------------------------------------------
+
+	private String extractDetail(BookmarkCommand.Create command, ContentDocument content) {
+		return truncate(command.wordIndex() == null ?
+				content.getScripts().get(Math.toIntExact(command.sentenceIndex())).getEnScript() :
+				content.getScripts().get(Math.toIntExact(command.sentenceIndex())).getEnScript()
+					.split(" ")[Math.toIntExact(command.wordIndex())],
+			255);
+	}
+
+	private Double extractStartTime(BookmarkCommand.Create command, ContentDocument content) {
+		if (contentRepository.findContentTypeById(command.contentId()).equals(ContentType.READING)) {
+			return null;
+		}
+
+		return ((YoutubeScript)content.getScripts()
+			.get(Math.toIntExact(command.sentenceIndex())))
+			.getStartTimeInSecond();
+	}
+
+	private String truncate(String content, int maxLength) {
+		return content.length() > maxLength ? content.substring(0, maxLength) : content;
 	}
 }
